refac(badges): moved badge parser to its own file
